old-tests/submission/Opera/web-messaging/message-channels/003.html

<!doctype html>
<title>onmessage implied start()</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function(t) {
  var channel = new MessageChannel();
  channel.port1.postMessage(1);
  var i = 0;
  channel.port2.onmessage = function() { i++; }; // implies start()
  setTimeout(t.step_func(function() { assert_equals(i, 1); t.done();}), 50);
});
</script>
